Create a standard package
