+++ comma/__init__.py
@@ -1,2 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
+
+__version__ = "0.2.0"
